+++ src/main/java/org/compiere/product/X_M_AttributeSetExclude.java
@@ -54,7 +54,7 @@ public class X_M_AttributeSetExclude extends PO implements I_M_AttributeSetExclu
      * @return Database Table information
      */
     public int getRowTableId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_AD_Table_ID);
+        Integer ii = getValue(COLUMNNAME_AD_Table_ID);
         if (ii == null) return 0;
         return ii;
     }
